The buttons along the bottom are:
^   General pointer - for moving (drag anywhere inside a box), resizing (bottom-right corner of a box, calling up the list of canvasses (double-clicking in a 'diagram' or 'display' box), and editing text.

diag  For drawing a model diagram box: drag a rectangle.

displ   For drawing a display box: ditto

txt   For drawing a text box: ditto

title   For drawing a title box (bigger font): ditto

>  <  Zoom in, out of poster canvas (but no scrolling, so you can't see bottom-right when zoomed in)

del   Delete a box (probably not implemented - can't remember)

save  load   For saving, loading a poster canvas file (basis for a template that can be re-used between sessions)

ps  Generate a postscript file

For some reason, the version I have at home (which I thought was the most recent) had a bug which meant that you couldn't call up the menu of canvasses by double-clicking - pretty fundamental!   Maybe I was in the middle of making some change when the phone rang (last May!), and I lost interest.   Anyway, I've made a change that at least gets around that, but God only knows what else is kicking around.   The programming for finding which canvasses to offer the user is very dodgy - probably the first place to start in tidying it up.

I mentioned on the phone the problem with text boxes.   Basically, running text that fits neatly into a box on screen produces lines that are too short, and with too big a line spacing, in the generated postscript.   That really does need to be sorted - the usefulness of this tool will be much less if you can't annotate the poster.

Another point is that this only works with canvasses.   Thus, if you do a 'plot value against time' helper, then you just see the gridded canvas where the graph is plotted - not the buttons, scales, etc.   This is another good reason to shift all displays to pure canvas operations (etc for buttons etc that really are for operating the helper).

I guess you won't be able to suss out all the features without me being around, so we can maybe have a look at it on Fri.

A screen dump of a sample poster (generated today, not a historical one, just to prove that it does still work) is attached.

Cheers,
Robert